[python/modules] Replace menu_creation.py with a dedicated package
This allows me to not have one giant module, which was becoming a bit too big for my taste.
